class.c: suppress a warning
* class.c (ins_methods_push): suppress a signed and unsigned
  comparison warning.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@50789 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ class.c
@@ -1075,7 +1075,7 @@ ins_methods_push(ID name, long type, VALUE ary, rb_method_visibility_t visi)
       case METHOD_VISI_PRIVATE:
       case METHOD_VISI_PROTECTED:
       case METHOD_VISI_PUBLIC:
-	visi = (type == visi);
+	visi = (type == (long)visi);
 	break;
       default:
 	visi = (type != METHOD_VISI_PRIVATE);
